requirements.txt
# PaisaKid / PaisaHiPaisa backend
Django>=5.0,<7
djangorestframework>=3.15
django-cors-headers>=4.6
Pillow>=10.0
